Bustabit_history/BustaAPI/requirements.txt
azure-core==1.7.0
azure-identity==1.2.0
azure-storage-blob==12.3.2
azure-storage-file-datalake==12.0.2
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==3.0
cycler==0.10.0
Django==2.2.4
djangorestframework==3.10.2
idna==2.10
isodate==0.6.0
joblib==0.16.0
kiwisolver==1.2.0
kneed==0.6.0
matplotlib==3.3.0
msal==1.4.2
msal-extensions==0.1.3
msrest==0.6.17
numpy==1.19.1
oauthlib==3.1.0
pandas==0.23.4
Pillow==7.2.0
portalocker==1.7.1
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.2
requests==2.24.0
requests-oauthlib==1.3.0
scikit-learn==0.23.1
scipy==1.5.2
six==1.15.0
sqlparse==0.3.0
threadpoolctl==2.1.0
urllib3==1.25.10
